# municipal-services/firenoc-services/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "eGov Fire NOC module.",
    "description": "APIs for Fire NOC module. This modules handles Below functionality.\n  1. Apply Fire NOC.\n  2. Search existing Fire NOC with different fiters.\n  3. Update exiting Fire NOC.\n",
    "contact": {
      "name": "eGovernments Foundation",
      "email": "contacts@egovernments.org"
    }
  },
  "schemes": [
    "https"
  ],
  "basePath": "/firenoc-services",
  "x-api-id": "org.egov.firenoc.services",
  "x-common-path": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml",
  "paths": {
    "/v1/_search": {
      "post": {
        "summary": "Get the list of Fire NOC(s) defined in the system.",
        "description": "1. Search and get Fire NOC(s)  based on defined search criteria.",
        "parameters": [
          {
            "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml#/parameters/requestInfo"
          },
          {
            "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml#/parameters/tenantId"
          },
          {
            "$ref": "#/parameters/pageNumber"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Search based on status.",
            "type": "string",
            "allowEmptyValue": true
          },
          {
            "name": "ids",
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "in": "query",
            "maxItems": 50,
            "description": "unique identifier of Fire NOC"
          },
          {
            "name": "applicationNumber",
            "in": "query",
            "description": "Unique application number for a Fire NOC application.",
            "type": "string",
            "minLength": 2,
            "maxLength": 64
          },
          {
            "name": "fireNOCNumber",
            "in": "query",
            "description": "The unique  license number  for a Fire NOC.",
            "type": "string",
            "minLength": 2,
            "maxLength": 64
          },
          {
            "name": "mobileNumber",
            "in": "query",
            "description": "The mobile number of a Fire NOC owner.",
            "type": "string",
            "pattern": "[0-9]"
          },
          {
            "name": "fireNOCType",
            "in": "query",
            "description": "Fire NOC type.",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fire NOC(s) Retrived Successfully",
            "schema": {
              "$ref": "#/definitions/FireNOCResponse"
            }
          },
          "400": {
            "description": "Invalid input.",
            "schema": {
              "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml#/definitions/ErrorRes"
            }
          }
        },
        "tags": [
          "Fire NOC"
        ]
      }
    },
    "/v1/_create": {
      "post": {
        "summary": "Create new Fire NOC.",
        "description": "To create new Fire NOC in the system. API supports bulk creation with max limit as defined in the Fire NOC Request. Please note that either whole batch succeeds or fails, there's no partial batch success. To create one Fire NOC, please pass array with one Fire NOC object.\n\n\nFollowing Conditions are applied -\n  1. Support Documents for a given Fire NOC will created. In case of not attaching the Mandatory Supported Documents creation of Fire NOC Fails.\n  2. Noc Number will be autogenerated on final approval for new License.\n  3. ApplicationNumber will be autogenerated for new Fire NOC\n",
        "parameters": [
          {
            "name": "FireNOCRequest",
            "in": "body",
            "description": "Details for the new Fire NOC(s) + RequestInfo meta data.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FireNOCRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ReponseInfo with Fire NOC(s) created successfully",
            "schema": {
              "$ref": "#/definitions/FireNOCResponse"
            }
          },
          "400": {
            "description": "Fire NOC(s) creation failed",
            "schema": {
              "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-0-0.yml#/definitions/ErrorRes"
            }
          }
        },
        "tags": [
          "Fire NOC"
        ]
      }
    },
    "/v1/_update": {
      "post": {
        "summary": "update existing Fire NOCs.",
        "description": "To update existing Fire NOC in the system. API supports bulk updation with max limit as defined in the Fire NOC Request. Please note that either whole batch succeeds or fails, there's no partial batch success. To update one Fire NOC, please pass array with one Fire NOC object.\n\nFollowing Conditions are applied -\n  1. Before payment Employee or Citizen can update any details of Fire NOC\n  2. After payment Citizen or Employee can't change any details of Fire NOC.\n",
        "parameters": [
          {
            "name": "FireNOCRequest",
            "in": "body",
            "description": "Details for the new Fire NOC(s) + RequestInfo meta data.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/FireNOCRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ReponseInfo with Fire NOC(s) updated sudcessfully",
            "schema": {
              "$ref": "#/definitions/FireNOCResponse"
            }
          },
          "400": {
            "description": "Fire NOC(s) updation failed",
            "schema": {
              "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-0-0.yml#/definitions/ErrorRes"
            }
          }
        },
        "tags": [
          "Fire NOC"
        ]
      }
    }
  },
  "definitions": {
    "Buildings": {
      "type": "object",
      "description": "It will contains building details",
      "properties": {
        "id": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "Unique Identifier of the Fire NOC building details (UUID)",
          "readOnly": true
        },
        "tenantId": {
          "type": "string",
          "description": "Unique Identifier of ULB",
          "maxLength": 128,
          "minLength": 2
        },
        "name": {
          "type": "string",
          "description": "name of the building"
        },
        "usageType": {
          "type": "string",
          "description": "building usage type"
        },
        "uoms": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BuildingUOM"
          }
        },
        "applicationDocuments": {
          "description": "1. List of all the required documents. 2. Application can be submitted without required document 3. Once all the document submitted then only application submition process will be completed. 4. Mandatry application documents for a fireNOC type and fireNOC subtype are defined under ApplicationDocument master which is defined under MDMS.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Document"
          }
        }
      },
      "required": [
        "tenantId",
        "usageType",
        "name",
        "uoms"
      ]
    },
    "PropertyDetails": {
      "type": "object",
      "description": "It will have fire noc related entities",
      "properties": {
        "id": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "Unique Identifier of the property details (UUID)",
          "readOnly": true
        },
        "propertyId": {
          "type": "string"
        },
        "address": {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml#/definitions/Address"
        }
      },
      "required": [
        "noOfBuildings",
        "buildingDetails"
      ]
    },
    "OwnerInfo": {
      "allOf": [
        {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/egov-user/contracts/v1-0-0.yml#/definitions/User"
        },
        {
          "type": "object",
          "properties": {
            "isPrimaryOwner": {
              "type": "boolean",
              "description": "The owner is primary or not"
            },
            "ownerShipPercentage": {
              "type": "number",
              "format": "double",
              "description": "Ownership percentage."
            },
            "ownerType": {
              "type": "string",
              "description": "Type of owner, based on this option Exemptions will be applied. This is master data defined in mdms.",
              "maxLength": 256,
              "minLength": 4
            },
            "relationship": {
              "type": "string",
              "description": "Relationship with owner.",
              "enum": [
                "FATHER",
                "HUSBAND"
              ]
            },
            "documents": {
              "description": "Document of the owner.",
              "items": {
                "$ref": "#/definitions/Document"
              }
            }
          }
        }
      ]
    },
    "Document": {
      "type": "object",
      "description": "A Object holds the basic data for a Fire NOC",
      "properties": {
        "tenantId": {
          "type": "string",
          "description": "Unique Identifier of ULB",
          "maxLength": 128,
          "minLength": 2
        },
        "documentType": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "1.Unique document type code from common master. 2. This is defined under mdms common master. 3. Object defination is defined under 'https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml#/definitions/DocumentType'"
        },
        "fileStoreId": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "Unique file store id of uploaded document."
        },
        "documentUid": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "document number(eg.Pan number, aadhar number)."
        },
        "auditDetails": {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-0-0.yml#/definitions/AuditDetails"
        }
      },
      "required": [
        "tenantId",
        "documentType",
        "fileStoreId"
      ]
    },
    "FireNOC": {
      "type": "object",
      "description": "A Object holds the basic data for a Fire NOC",
      "properties": {
        "id": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "Unique Identifier of the Fire NOC (UUID)",
          "readOnly": true
        },
        "tenantId": {
          "type": "string",
          "description": "Unique Identifier of ULB",
          "maxLength": 128,
          "minLength": 2
        },
        "fireNOCNumber": {
          "type": "string",
          "description": "Unique Fire NOC Number of the Fire NOC. This is  unique in system for a tenant. This is mandatory but always be generated on the final approval.",
          "maxLength": 64,
          "minLength": 2,
          "readOnly": true
        },
        "provisionFireNOCNumber": {
          "type": "string",
          "description": "Unique Fire NOC Number of the Provision Fire NOC number that will be used for linking provision fire NOC number with new fire NOC number. This is  unique in system for a tenant. This is mandatory but always be generated on the final approval.",
          "maxLength": 64,
          "minLength": 2
        },
        "oldFireNOCNumber": {
          "type": "string",
          "description": "Unique Old License Number of the Fire NOC. This is  unique in system for a tenant. This is mandatory  for legacy license(DataEntry).",
          "maxLength": 64,
          "minLength": 2
        },
        "dateOfApplied": {
          "type": "integer",
          "format": "int64",
          "description": "Applied Date of the fire NOC as epoch"
        },
        "fireNOCDetails": {
          "$ref": "#/definitions/FireNOCDetails"
        },
        "auditDetails": {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-0-0.yml#/definitions/AuditDetails"
        }
      },
      "required": [
        "tenantId",
        "fireNOCType",
        "dateOfApplied",
        "fireNOCDetails"
      ]
    },
    "FireNOCDetails": {
      "type": "object",
      "description": "A Object holds the basic data for a Fire NOC",
      "properties": {
        "id": {
          "type": "string",
          "minLength": 2,
          "maxLength": 64,
          "description": "Unique Identifier of the Fire FireNOC Details (UUID)",
          "readOnly": true
        },
        "applicationNumber": {
          "type": "string",
          "description": "Unique Application FireNOC Number of the Fire FireNOC. This is  unique in system for a tenant. This is mandatory but always be generated on the final approval.",
          "maxLength": 64,
          "minLength": 2,
          "readOnly": true
        },
        "fireNOCType": {
          "type": "string",
          "description": "type of fire NOC from mdms"
        },
        "firestationId": {
          "type": "string",
          "description": "Fire station id where we are applying fire fireNOC."
        },
        "applicationDate": {
          "type": "integer",
          "format": "int64",
          "description": "date on which applicaiton has been generated for new Fire NOC.",
          "readOnly": true
        },
        "financialYear": {
          "type": "string",
          "description": "Fire NOC applicable for financial year.",
          "maxLength": 64,
          "minLength": 2
        },
        "issuedDate": {
          "type": "integer",
          "format": "int64",
          "description": "1. License issued Date of the Fire NOC as epoch. 2. Application approval date.",
          "readOnly": true
        },
        "validFrom": {
          "type": "integer",
          "format": "int64",
          "description": "Date from when Fire NOC is valid as epoch"
        },
        "validTo": {
          "type": "integer",
          "format": "int64",
          "description": "Expiry Date of the Fire NOC as epoch"
        },
        "action": {
          "type": "string",
          "description": "1. Perform action to change the state of the Fire NOC. 2. INITIATE, if application is getting submitted without required document. 3. APPLY, if application is getting submitted with application documents, in that case api will validate all the required application document. 4. APPROVE action is only applicable for specific role, that role has to be configurable at service level. Employee can approve a application only if application is in APPLIED state and Fire NOC fees is paid.",
          "enum": [
            "INITIATE",
            "APPLY",
            "APPROVE",
            "REJECT",
            "CANCEL"
          ]
        },
        "channel": {
          "type": "string",
          "description": "License can be created from different channels",
          "maxLength": 64,
          "minLength": 2,
          "enum": [
            "COUNTER",
            "CITIZEN",
            "DATAENTRY"
          ]
        },
        "noOfBuildings": {
          "type": "string",
          "description": "it might be single or multiple"
        },
        "buildings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Buildings"
          }
        },
        "propertyDetails": {
          "$ref": "#/definitions/PropertyDetails"
        },
        "applicantDetails": {
          "type": "object",
          "description": "This will have details about applicant details",
          "properties": {
            "ownerShipType": {
              "description": "type pf the owner ship",
              "type": "string"
            },
            "owners": {
              "description": "Fire NOC owners, these will be citizen users in system.",
              "type": "array",
              "items": {
                "$ref": "#/definitions/OwnerInfo"
              }
            },
            "additionalDetail": {
              "type": "object",
              "description": "Json object to store additional details about license, this will be used when ownership is intitution or others"
            }
          }
        },
        "additionalDetail": {
          "type": "object",
          "description": "Json object to store additional details about license"
        },
        "auditDetails": {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-0-0.yml#/definitions/AuditDetails"
        }
      },
      "required": [
        "tenantId",
        "fireNOCType",
        "financialYear",
        "propertyDetails",
        "applicantDetails",
        "action",
        "channel",
        "firestationId"
      ]
    },
    "FireNOCRequest": {
      "description": "Contract class to receive request. Array of Noc items are used",
      "properties": {
        "RequestInfo": {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-1-1.yml#/definitions/RequestInfo"
        },
        "FireNOCs": {
          "description": "Used for search result and create only",
          "type": "array",
          "minimum": 1,
          "maximum": 100,
          "items": {
            "$ref": "#/definitions/FireNOC"
          }
        }
      },
      "required": [
        "RequestInfo",
        "FireNOCs"
      ]
    },
    "FireNOCResponse": {
      "description": "Contract class to send response. Array of Fire NOC items are used in case of search results or response for create, whereas single Fire NOC item is used for update",
      "properties": {
        "ResponseInfo": {
          "$ref": "https://raw.githubusercontent.com/egovernments/egov-services/master/docs/common/contracts/v1-0-0.yml#/definitions/ResponseInfo"
        },
        "FireNOCs": {
          "description": "Used for search result and create only",
          "type": "array",
          "maximum": 500,
          "items": {
            "$ref": "#/definitions/FireNOC"
          }
        }
      }
    },
    "BuildingType": {
      "type": "object",
      "description": "This is mdms defination for BuildingType object",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "UoMs": {
          "description": "List of UoMs based building type",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "FireNOCType": {
      "type": "object",
      "description": "Type of FireNOC Provisional and New",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "OwnershipType": {
      "type": "object",
      "description": "Owner ship type Individual/multiple/Insti",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "OwnerSpecialCategory": {
      "type": "object",
      "description": "Owner special category Widow/freedom fighter",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "FireStations": {
      "type": "object",
      "description": "Fire staion where we apply for fire NOC",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "baseTenantId": {
          "type": "string",
          "description": "When multiple city sharing single firestation that time we can mention base tenant id for the firestation."
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "FireStationsToULBMapping": {
      "type": "object",
      "description": "Fire staion to ulb mapping",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "fireStationCode": {
          "type": "string",
          "description": "Unique value reffered to FireStations master"
        },
        "boundaryCode": {
          "type": "string",
          "description": "Unique value reffered to boundary master, it will be dot seperated eg- pb.amritsar.SC4"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "FireNOCMetaInfo": {
      "type": "object",
      "description": "This master used to maintain metadata of fire fireNOC",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "firestationMapping": {
          "type": "string",
          "description": "Here we will mention whether its a fire to ulb mapping or fire to mohalla mapping, depends on this configuration we will filter firestation to ulb mapping."
        },
        "certificateValidity": {
          "type": "string",
          "description": "Here we will mention validity of mode eg- days/months/years."
        },
        "certificateEndOfValidity": {
          "type": "string",
          "description": "Here we will mention end of validity eg- calendar/financial year."
        },
        "overallSLA": {
          "type": "number",
          "description": "It will contain overall SLA for process"
        }
      }
    },
    "UOM": {
      "type": "object",
      "description": "This master will have list of UOM's",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "BuildingUOM": {
      "type": "object",
      "description": "This master will have list of UOM's",
      "properties": {
        "code": {
          "type": "string",
          "description": "Code of the unit code"
        },
        "value": {
          "type": "string",
          "description": "Value entered for the uom"
        },
        "isActiveUom": {
          "type": "boolean",
          "description": "Active uom for current usage type"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "Status": {
      "type": "object",
      "description": "This master will have list of Status",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "PropertyType": {
      "type": "object",
      "description": "This master will have list of property type Single/Multiple building",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "active": {
          "type": "boolean"
        }
      }
    },
    "SLA": {
      "type": "object",
      "description": "This master will have mapping between Status to SLA",
      "properties": {
        "code": {
          "type": "string",
          "description": "Unique value for each entry in mdms data"
        },
        "statusCode": {
          "type": "string",
          "description": "This is will be referene to status master"
        },
        "noOfDays": {
          "type": "string",
          "description": "It will have status completion days."
        },
        "active": {
          "type": "boolean"
        }
      }
    }
  },
  "parameters": {
    "pageSize": {
      "name": "pageSize",
      "in": "query",
      "description": "Number of records returned.",
      "type": "integer",
      "minimum": 0,
      "exclusiveMinimum": true,
      "maximum": 100,
      "exclusiveMaximum": false,
      "multipleOf": 10,
      "default": 20
    },
    "pageNumber": {
      "name": "pageNumber",
      "in": "query",
      "description": "Page number",
      "type": "integer",
      "default": 1
    },
    "sortProperty": {
      "name": "sort",
      "in": "query",
      "description": "Fire NOC results will be sorted by fireNOCNumber ascending by default if this parameter is not provided.",
      "type": "array",
      "uniqueItems": true,
      "minItems": 1,
      "maxItems": 3,
      "collectionFormat": "pipes",
      "items": {
        "type": "string",
        "pattern": "[-+](fireNOCNumber|fireNOCType|ownerName)"
      },
      "default": [
        "+fireNOCNumber"
      ]
    }
  },
  "tags": [
    {
      "name": "Fire NOC",
      "description": "APIs required for Fire NOC."
    }
  ]
}